Fixed rather nasty alignment bug.
On my gcc arm, 9.3, with wanrings turned on, there were many warnings about

casting from a smaller size to a larger size.  I ignored them.

That was a bad idea.  It seems that at least on the architecture I'm using
ARM cortext M7, you really can't cast from a uint8_t to a uint32_t alignment
It caused a hard-to-debug error that took all day to find because I didn't have
debugging on my target board. (A teensy 4.1).  I thought my problem was
in the USB library.

This has the internal buffer be an int32 instead of int8 type, which
keeps all the casts aligned on 32-bit boundaries, and keeps the
compiler happy.

Also, fixed a few unsigned int for loop wanrings while I was at it.

When allocating, now all sizes are increased to be a multiple of 4 larger
than the size in bytes.

This fixes the crashing I was seeing.

This only showed up on -O3, and never on -O0.